+++ examples/pre-signed-url/data.tf
@@ -1,0 +1,3 @@
+data "aws_region" "current_region" {} # Find region, e.g. us-east-1
+
+data "aws_caller_identity" "this" {}